% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{baseline}
\alias{baseline}
\title{calculating baseline hazard using the result from a \code{coxtv} or \code{coxtp} object}
\usage{
baseline(fit, ...)
}
\arguments{
\item{fit}{model from \code{coxtv} or \code{coxtp}}
}
\value{
a list with three components
\item{time}{the unique observation time}
\item{hazard}{the baseline hazard corresponding to each unqiue time point}
\item{cumulHaz}{the cumulative baseline hazard corresponding to each unqiue time point}
}
\description{
The baseline estimation is the baseline hazard at each observation time when holding all the covariates equals to zero.
}
\examples{
\dontrun{
data(ExampleData)
z <- ExampleData$x
time  <- ExampleData$time
event <- ExampleData$event

fit   <- coxtv(event = event, z = z, time = time)
base.est <- baseline(fit)
}


}
