% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtp.base.R
\name{confint.coxtp}
\alias{confint.coxtp}
\title{get confidence interval from a 'coxtp' object}
\usage{
\method{confint}{coxtp}(fit, times, parm, level = 0.95)
}
\arguments{
\item{fit}{fitted \code{"coxtp"} model}

\item{times}{the time interval to be estamtied. The default value is the time of the fitted model}

\item{parm}{the names of parameter}

\item{level}{the confidence level. Default is 0.95.}
}
\description{
get confidence interval from a 'coxtp' object
}
\examples{
data(ExampleData)
z <- ExampleData$x
time <- ExampleData$time
event <- ExampleData$event
fit <- coxtp(event = event, z = z, time = time)
confit(fit)

}
