% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtp.base.R, R/coxtv.base.R
\name{confint.coxtp}
\alias{confint.coxtp}
\alias{confint.coxtv}
\title{get confidence interval from a fitted object}
\usage{
\method{confint}{coxtp}(fit, times, parm, level = 0.95)

\method{confint}{coxtv}(fit, times, parm, level = 0.95, ...)
}
\arguments{
\item{fit}{fitted \code{"coxtv"} model}

\item{times}{the time interval to be estamtied. The default value is the time of the fitted model}

\item{parm}{the names of parameter}

\item{level}{the confidence level. Default is 0.95.}
}
\description{
get confidence interval from a fitted object
}
\examples{
\dontrun{
data(ExampleData)
z <- ExampleData$x
time <- ExampleData$time
event <- ExampleData$event
fit <- coxtp(event = event, z = z, time = time)
confint(fit$lambda1)
}

\dontrun{
data(ExampleData)
z <- ExampleData$x
time <- ExampleData$time
event <- ExampleData$event
fit <- coxtv(event = event, z = z, time = time)
confint(fit)
}

}
