% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtv.baseline.R
\name{coxtv.baseline}
\alias{coxtv.baseline}
\title{calculating baseline hazard using the result from a \code{coxtv} object}
\usage{
coxtv.baseline(fit, ...)
}
\arguments{
\item{fit}{fitted model from \code{coxtv}}
}
\value{
a list with components
\item{time}{the unique event time points}
\item{hazard}{the baseline hazard corresponding to each unqiue time point}
\item{cumulHaz}{the cumulative baseline hazard corresponding to each unqiue time point}
}
\description{
calculating baseline hazard using the result from a \code{coxtv} object
}
\examples{
data("ExampleData")
z <- ExampleData$x
time <- ExampleData$time
event <- ExampleData$event
fit <- coxtv(event = event, z = z, time = time)
base.est = coxtv.baseline(fit)
}
