% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleData.R
\docType{data}
\name{ExampleData}
\alias{ExampleData}
\title{Example Dataset}
\format{
A data.frame wit
\describe{
\item{V1}{Simulated covariate V1, Continuous variable. True time-dependent function is b(t)=1}
\item{V2}{Simulated covariate V2, Continuous variable. True time-dependent function is b(t)=sin(3\emph{pi}x/4)}
\item{event}{Simulated event variable, Binary varible with 0, 1 }
\item{time}{Simulated time variable, Continous variable with non-negative value}
}
}
\usage{
ExampleData
}
\description{
Example data for surtvep with 4000 observations of 2 variables:
}
\details{
A simulated data set containing two 2 variables.
}
\keyword{datasets}
