% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ExampleDataBinary}
\alias{ExampleDataBinary}
\title{ExampleDataBinary
Example data for surtvep with 2000 observations of 2 binary variables:}
\format{
A List containing the following elements:
\describe{
\item{V1}{Simulated covariate V1, Binary variable. True time-dependent function is b(t)=1}
\item{V2}{Simulated covariate V2, Binary variable. True time-dependent function is b(t)=exp(-1.5*t)}
\item{event}{Simulated event variable, Binary varible with 0, 1 }
\item{time}{Simulated time variable, Continous variable with non-negative value}
}
}
\usage{
data(ExampleDataBinary)
}
\description{
A simulated data set containing two 2 variables.
}
\keyword{datasets}
