% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC.R
\name{IC}
\alias{IC}
\title{calculating information criteria from a \code{coxtp} object}
\usage{
IC(fit, IC.prox, ...)
}
\arguments{
\item{fit}{model from \code{coxtp}.}

\item{IC.prox}{when calculating information criteria, there might be numerical issues (e.g. the Hessian matrix is close to be singular).
In such cases warnings will be given.
If \code{IC.prox = true}, we modified the diagonal of the Hessian matrix a bit, which can lead to more stable estimates.
Default is \code{FALSE}.}
}
\value{
\item{model.mAIC}{an object with S3 class \code{"coxtp"} using mAIC to select the tuning parameter.}
\item{model.TIC}{an object with S3 class \code{"coxtp"} using TIC to select the tuning parameter.}
\item{model.GIC}{an object with S3 class \code{"coxtp"} using GIC to select the tuning parameter.}
\item{mAIC}{a sequence of mAIC values for the different tuning parameters \code{lambda} from \code{"coxtp"}.}
\item{TIC}{a sequence of TIC values for the different tuning parameters \code{lambda} from \code{"coxtp"}.}
\item{GIC}{a sequence of GIC values for the different tuning parameters \code{lambda} from \code{"coxtp"}.}
}
\description{
This function is to calculate information criteria from a \code{coxtp} object to select the penalization tuning parameter.
}
\details{
In order to select the proper smoothing parameter, we utilize the idea of information criteria.
We provide four different information criteria to select the optimal smoothing parameter \eqn{\lambda}.
Generally, mAIC, TIC and GIC select similar parameters and the difference of resulting estimates are barely noticeable.
See details in Luo et al. (2022).
}
\examples{
data(ExampleData)
z <- ExampleData$x
time <- ExampleData$time
event <- ExampleData$event
fit <- coxtp(event = event, z = z, time = time)
IC  <- IC(fit)


}
