% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtp.plot.R
\name{plot.ic.coxtp}
\alias{plot.ic.coxtp}
\title{Plotting result from a \code{ic.coxtp} object}
\usage{
\method{plot}{ic.coxtp}(
  model,
  IC = "AIC",
  times,
  CI = TRUE,
  level = 0.95,
  exponentiate = FALSE,
  xlab,
  ylab,
  xlim,
  ylim,
  save = FALSE,
  allinone = FALSE,
  title,
  linetype,
  fill,
  color,
  labels,
  expand,
  ...
)
}
\arguments{
\item{IC}{The Criteria selected for the plot. Default is "AIC", which uses AIC to select the tunning parameter.}

\item{times}{The time scale for the plot}

\item{CI}{This argument controls the confidence interval on the plot. With 'CI' = TRUE, the confidence interval will be plotted.}

\item{level}{The confidence level.}

\item{fit}{Fitted "coxtp" model object from the main function "coxtp".}
}
\description{
Plotting result from a \code{ic.coxtp} object
}
\examples{
data("ExampleData")
z     <- ExampleData$x
time  <- ExampleData$time
event <- ExampleData$event
model   <- coxtp(event = event, z = z, time = time, lambda_spline = c(1))
plot(model, IC = "AIC", allinone = TRUE)

}
