% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtv.plot.R
\name{plot.surtiver}
\alias{plot.surtiver}
\title{Plotting result from coxtp() function}
\usage{
\method{plot}{surtiver}(
  fit,
  times,
  parm,
  CI = TRUE,
  level = 0.95,
  exponentiate = FALSE,
  xlab,
  ylab,
  xlim,
  ylim,
  save = FALSE,
  allinone = FALSE,
  title,
  linetype,
  fill,
  color,
  labels,
  expand,
  ...
)
}
\arguments{
\item{fit}{Model get from coxtp}

\item{IC}{The Creteria selected for the plotting}

\item{coef}{The variable that needed to be plotted}
}
\description{
Plotting result from coxtp() function
}
