% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{Example data for surtvep with 5000 observations of 4 variables:}
\format{
A data.frame
\describe{
\item{V1}{Simulated covariate V1, Binary variable with 0, 1. True time-dependent function is b(t)=1}
\item{V2}{Simulated covariate V2, Binary variable with 0, 1. True time-dependent function is b(t)=exp(-1.5*t)}
\item{event}{Simulated event variable, Binary varible with 0, 1 }
\item{time}{Simulated time variable, Continous variable with non-negative value}
}
}
\usage{
sim_data
}
\description{
Example data for surtvep with 5000 observations of 4 variables:
}
\keyword{datasets}
