% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulN2kOP2}
\alias{simulN2kOP2}
\title{Example data for surtvep}
\format{
A data.frame with 2000 observations of 4 variables:
\describe{
\item{z.1}{Description?}
\item{z.2}{Description?}
\item{delta}{Description?}
\item{time}{Description?}
}
}
\usage{
simulN2kOP2
}
\description{
Example data for surtvep
}
\keyword{datasets}
