% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{StrataExample}
\alias{StrataExample}
\title{example data for stratified model illustration}
\format{
A list containing the following elements:
\describe{
\item{x}{simulated binary covariate V1 and V2. True time-independent function is b(t)=1 and true time-varying function is b(t)=sin(3tpi/4).}
\item{event}{simulated event variable, binary variable with 0, 1.}
\item{time}{simulated time variable, continuous variable with non-negative value.}
\item{strata}{simulated strata variable, patients in different stratums have different baseline hazards.}
}
}
\usage{
data(StrataExample)
}
\description{
A simulated data set containing two 2 variables.
}
\keyword{datasets}
