% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surtvp.R
\name{surtvep}
\alias{surtvep}
\title{Title}
\usage{
surtvep(
  event = delta,
  z = z,
  time = time,
  spline = "Smooth-spline",
  nsplines = 8,
  ties = "Breslow",
  tol = 1e-09,
  iter.max = 20L,
  method = "Newton",
  lambda = 1e+08,
  btr = "static",
  tau = 0.5,
  stop = "ratch",
  parallel = FALSE,
  threads = 1L,
  degree = 3L,
  TIC = FALSE,
  TIC_prox = FALSE,
  lambda_spline = 0,
  ord = 4,
  fixedstep = FALSE,
  ICLastOnly = TRUE
)
}
\arguments{
\item{event}{indicator for occurrence of event (1 if event, 0 else)}

\item{z}{define...}

\item{time}{= time to event}

\item{spline}{define...= "Smooth-spline",}

\item{nsplines}{define...= 8,}

\item{ties}{define...= "none",}

\item{tol}{define...= 1e-6,}

\item{iter.max}{define...= 20L,}

\item{method}{define...= "Newton",}

\item{btr}{define...= "dynamic",}

\item{stop}{define...= "ratch",}

\item{parallel}{define...= TRUE,}

\item{threads}{define...= 3L,}

\item{degree}{define...= 3L,}

\item{lambda_spline}{define...= lambda_spline,}

\item{ICLastOnly}{define...= TRUE}
}
\value{
A list containing ....
}
\description{
Description
}
\examples{

# load example data
data(simulN2kOP2)

z = simulN2kOP2[, c("z.1", "z.2")]
delta = simulN2kOP2$delta
time = simulN2kOP2$time

# specify smoothing parameter lambda
lambda_spline = c(1:5)

# model fitting
models = surtvep(event = delta, z = z, time = time, 
                  lambda_spline = lambda_spline,
                  spline="Smooth-spline", nsplines=8, ties="none", 
                  tol=1e-6, iter.max=20L, method="Newton",
                  btr="dynamic", stop="ratch", 
                  parallel=TRUE, threads=3L, degree=3L,
                  ICLastOnly = TRUE)

}
