% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surtvier.plot.R
\name{surtvier.plot}
\alias{surtvier.plot}
\title{Plotting result from surtvier() function}
\usage{
surtvier.plot(fit, coef)
}
\arguments{
\item{fit}{Model get from coxtp}

\item{coef}{The variable that needed to be plotted}

\item{IC}{The Creteria selected for the plotting}
}
\description{
Plotting result from surtvier() function
}
