% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvef.ph.R
\name{tvef.ph}
\alias{tvef.ph}
\title{test the proportional hazards assumption from a \code{coxtv} or \code{coxtp} object}
\usage{
tvef.ph(fit, parm)
}
\arguments{
\item{fit}{fitted \code{"coxtv"} or \code{"coxtp"}  model}

\item{parm}{the names of parameter to be tested}
}
\description{
test the proportional hazards assumption from a \code{coxtv} or \code{coxtp} object
}
\examples{
data(ExampleData)
z <- ExampleData$x
time <- ExampleData$time
event <- ExampleData$event
fit <- coxtv(event = event, z = z, time = time)
tvef.ph(fit)

}
