% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvef.ph.R
\name{tvef.zero.time}
\alias{tvef.zero.time}
\title{test the significance of the covariates from a \code{coxtv} or \code{coxtp} object using a Wald test statistic}
\usage{
tvef.zero.time(fit, times, parm)
}
\arguments{
\item{fit}{fitted \code{"coxtv"} or \code{"coxtp"}  model}

\item{times}{the time points to test if the covariate is significant or not}

\item{parm}{the names of parameter to be tested}
}
\description{
test the significance of the covariates at each time point.
}
\examples{
data(ExampleData)
z <- ExampleData$x
time  <- ExampleData$time
event <- ExampleData$event
fit   <- coxtv(event = event, z = z, time = time)
test  <- tvef.zero.time(fit)

}
\seealso{
\code{\link{tvef.ph}} \code{\link{tvef.zero}}
}
