% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtv.base.R
\name{confint.coxtv}
\alias{confint.coxtv}
\title{get confidence interval from a 'coxtv' object}
\usage{
\method{confint}{coxtv}(fit, times, parm, level = 0.95, ...)
}
\arguments{
\item{fit}{fitted \code{"coxtv"} model}

\item{times}{the time interval to be estamtied. The default value is the time of the fitted model}

\item{parm}{the names of parameter}

\item{level}{the confidence level. Default is 0.95.}
}
\description{
get confidence interval from a 'coxtv' object
}
\examples{
data(ExampleData)
z <- ExampleData$x
time <- ExampleData$time
event <- ExampleData$event
fit <- coxtv(event = event, z = z, time = time)
confit(fit)

}
