% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtp.base.R, R/coxtv.base.R
\name{confint.coxtp}
\alias{confint.coxtp}
\alias{confint.coxtv}
\title{get confidence intervals of time-varying coefficients from a fitted object}
\usage{
\method{confint}{coxtp}(fit, times, parm, level = 0.95)

\method{confint}{coxtv}(fit, times, parm, level = 0.95, ...)
}
\arguments{
\item{fit}{fitted \code{"coxtv"} model.}

\item{times}{the time points for which the confidence intervals to be estimated.
The default value is the unique observed event times in the dataset fitting the time-varying effects model.}

\item{parm}{the names of parameters.}

\item{level}{the confidence level. Default is 0.95.}
}
\description{
Get confidence intervals of time-varying coefficients from a fitted \code{coxtv} or \code{coxtp} object.
}
\examples{
\dontrun{
data(ExampleData)
z <- ExampleData$x
time <- ExampleData$time
event <- ExampleData$event
fit <- coxtp(event = event, z = z, time = time)
confint(fit$lambda1)
}

\dontrun{
data(ExampleData)
z <- ExampleData$x
time <- ExampleData$time
event <- ExampleData$event
fit <- coxtv(event = event, z = z, time = time)
confint(fit)
}

}
