% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtp.baseline.R
\name{coxtp.baseline}
\alias{coxtp.baseline}
\title{calculating baseline hazard using the result from a \code{coxtv} object}
\usage{
coxtp.baseline(fit, delta, z, time, strata = c())
}
\arguments{
\item{fit}{model from \code{coxtp}}
}
\value{
a list with three components
\item{time}{the unique observation time}
\item{hazard}{the baseline hazard corresponding to each unqiue time point}
\item{cumulHaz}{the cumulative baseline hazard corresponding to each unqiue time point}
}
\description{
The baseline estimation is the baseline hazard at each observation time when holding all the covariates equals to zero.
}
