% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtp.baseline.R
\name{coxtp.baseline}
\alias{coxtp.baseline}
\title{calculating Baseline hazard using the result from coxtp()}
\usage{
coxtp.baseline(fit, delta, z, time, strata = c())
}
\arguments{
\item{fit}{model get from \code{coxtp} or \code{coxtv}}

\item{delta}{event vector, should be a vector containing 0 or 1}

\item{z}{Covariate matrix}

\item{time}{Time vector, should be a vector with non-negative numeric value}

\item{strata}{stratification group defined in the data. If there exist stratification group, please enter as vector.}
}
\description{
calculating Baseline hazard using the result from coxtp()
}
