% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtp.plot.R
\name{coxtp.plot}
\alias{coxtp.plot}
\title{Title}
\usage{
coxtp.plot(fit, IC = "AIC", coef)
}
\arguments{
\item{fit}{Model get from coxtp}

\item{IC}{The Creteria selected for the plotting}

\item{coef}{The variable that needed to be plotted}
}
\value{

}
\description{
Title
}
