% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtp.predict.R
\name{coxtp.predict}
\alias{coxtp.predict}
\title{Title}
\usage{
coxtp.predict(fit, baseline, newdata = c(), strata = FALSE)
}
\arguments{
\item{fit}{Model get from coxtp}

\item{baseline}{Baseline estimation from coxtp.baseline or the arbitary baseline entered}

\item{newdata}{New data in vector}

\item{strata}{Whether there is stratification in the dataset, default is FALSE.}
}
\value{

}
\description{
Title
}
