% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtp.predict.R
\name{coxtp.predict}
\alias{coxtp.predict}
\title{make predictions from a "coxtv" or "coxtp" object.}
\usage{
coxtp.predict(fit, baseline, newdata = c(), strata = FALSE, out_seq = NULL)
}
\arguments{
\item{fit}{Model get from coxtp}

\item{baseline}{Baseline estimation from coxtp.baseline or the arbitary baseline entered}

\item{newdata}{New data in vector}

\item{strata}{Whether there is stratification in the dataset, default is FALSE.}

\item{out_seq}{Output time sequence. Default is out_seq=NULL. When default, will output the time sequence according to input fit}
}
\description{
Similar to other predict methods, this functions predicts fitted values, logits, coefficients and more from a fitted ""coxtv" or "coxtp"object.
}
