% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtp.R
\name{coxtp}
\alias{coxtp}
\title{Title}
\usage{
coxtp(
  event,
  z,
  time,
  spline = "Smooth-spline",
  nsplines = 8,
  ties = "Breslow",
  tol = 1e-09,
  iter.max = 20L,
  method = "Newton",
  lambda = 1e+08,
  btr = "static",
  tau = 0.5,
  stop = "ratch",
  parallel = FALSE,
  threads = 1L,
  degree = 3L,
  TIC = FALSE,
  TIC_prox = FALSE,
  lambda_spline = 0,
  ord = 4,
  fixedstep = FALSE,
  ICLastOnly = TRUE,
  penalizestop = FALSE
)
}
\arguments{
\item{event}{}

\item{z}{}

\item{time}{}

\item{spline}{}

\item{nsplines}{}

\item{ties}{}

\item{tol}{}

\item{iter.max}{}

\item{method}{}

\item{lambda}{}

\item{btr}{}

\item{tau}{}

\item{stop}{}

\item{parallel}{}

\item{threads}{}

\item{degree}{}

\item{TIC}{}

\item{TIC_prox}{}

\item{lambda_spline}{}

\item{ord}{}

\item{fixedstep}{}

\item{ICLastOnly}{}

\item{penalizestop}{}
}
\value{

}
\description{
Title
}
