% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtv.baseline.R
\name{coxtv.baseline}
\alias{coxtv.baseline}
\title{Calculating Baseline hazard using the result from a \code{coxtv} object}
\usage{
coxtv.baseline(fit, event, z, time, strata = NULL, ...)
}
\arguments{
\item{fit}{Model get from coxtp}

\item{z}{Covariate matrix}

\item{time}{Time vector, should be a vector with non-negative numeric value}

\item{strata}{stratification group defined in the data. If there exist stratification group, please enter as vector.}

\item{delta}{event vector, should be a vector containing 0 or 1}
}
\value{
a list with components
\item{time}{the unique event time points}
\item{hazard}{the baseline hazard corresponding to each unqiue time point}
\item{cumulHaz}{the cumulative baseline hazard corresponding to each unqiue time point}
}
\description{
Calculating Baseline hazard using the result from a \code{coxtv} object
}
\examples{
data("ExampleDataBinary")
z <- ExampleDataBinary$x
time <- ExampleDataBinary$time
event <- ExampleDataBinary$event
fit <- coxtv(event = event, z = z, time = time)
base.est = coxtv.baseline(fit, event, z, time)
}
