% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtv.base.R
\name{coxtv}
\alias{coxtv}
\title{fit a cox Non-proportional Hazards model:}
\usage{
coxtv(
  event,
  z,
  time,
  strata = NULL,
  spline = "P-spline",
  nsplines = 8,
  ties = "Breslow",
  control,
  ...
)
}
\arguments{
\item{event}{event vector, should be a vector containing 0 or 1}

\item{z}{Covariate matrix}

\item{time}{Time vector, should be a vector with non-negative numeric value}

\item{strata}{stratification group defined in the data. If there exist stratification group, please enter as vector.}

\item{spline}{The spline term for Penalized Newton's Method(Add section
Number related to the paper). Default setting is \strong{\code{spline="Smooth-spline"}}}

\item{nsplines}{Number of base functions in the B-splines, default is 8.}

\item{ties}{Ways to deal with ties, default is \strong{\code{ties="Breslow"}}:}

\item{tol}{Convergence threshold. The default threshold is set as \strong{\code{tol=1e-6}}}

\item{iter.max}{Maximum Iteration number, default is \strong{\code{iter.max=20L}}}

\item{method}{Selecting Method used, default is \strong{\code{method="Newton"}}}

\item{lambda}{Parameter for Proximal Newton's Method. Default is \strong{\code{lambda=1e8}}}

\item{btr}{Backtracking line search approach, default is \code{btr="static"}:}

\item{tau}{(Alpha or beta?) in the Newton's Method, Default is \strong{\code{tau=0.5}}. Used to control for step size.}

\item{stop}{Stopping rule, default is \strong{\code{stop="ratch"}}:}

\item{parallel}{Parallel computation, Default is \strong{\code{parallel=FALSE}}}

\item{threads}{Parallel computation parameter(number of cores)Default is \strong{\code{threads=1L}}}

\item{degree}{Degree of smoothing spline. Default setting is \strong{\code{degree=3L}}.}

\item{ord}{Specify which derivative to penalize. Default setting is \strong{\code{ord=4}}.}

\item{fixedstep}{There might be times when the stopping criteria not working, thus,
the number of steps could be set manually. Default value is \strong{\code{fixedstep = FALSE}}, if it is true, will stop by \code{iter.max}}
}
\value{
An object with S3 class \code{"coxtv"}.
\item{ctrl.pts}{the estimated covariate coefficient}
\item{theta.list}{the estimated covariate function across all iterations}
\item{VarianceMatrix}{The variance matrix of the estimated function}
\item{internal.knots}{The internal knot locations}
\item{info}{The second order derivative}
}
\description{
Fit a cox Non-proportional Hazards model via maximum likelihood.
}
\details{
put the link to the wesite()...
}
\examples{
data(ExampleData)
z <- ExampleData$x
time <- ExampleData$time
event <- ExampleData$event
fit <- coxtv(event = event, z = z, time = time)

}
\seealso{
\code{coef}, \code{plot}, \code{coef} and \code{plot} methods,
and the \code{coxtp} function.
}
