% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.coxtp.r
\name{cv.coxtp}
\alias{cv.coxtp}
\title{fit a cox Non-proportional Hazards model with P-spline or Smoothing-spline, penalization coefficient is provided by cross validation}
\usage{
cv.coxtp(
  event,
  z,
  time,
  strata = c(),
  spline = "Smooth-spline",
  nsplines = 8,
  ties = "Breslow",
  tol = 1e-09,
  iter.max = 20L,
  method = "ProxN",
  lambda = 1e+08,
  btr = "dynamic",
  tau = 0.5,
  stop = "ratch",
  parallel = FALSE,
  threads = 1L,
  degree = 3L,
  TIC_prox = FALSE,
  lambda.spline = 0,
  ord = 4,
  fixedstep = FALSE
)
}
\arguments{
\item{event}{event vector, should be a vector containing 0 or 1}

\item{z}{Covariate matrix}

\item{time}{Time vector, should be a vector with non-negative numeric value}

\item{strata}{stratification group defined in the data. If there exist stratification group, please enter as vector.}

\item{spline}{The spline term for Penalized Newton's Method(Add section
Number related to the paper). Default setting is \strong{\code{spline="Smooth-spline"}}}

\item{nsplines}{Number of base functions in the B-splines, default is 8.}

\item{ties}{Ways to deal with ties, default is \strong{\code{ties="Breslow"}}:}

\item{tol}{Convergence threshold. The default threshold is set as \strong{\code{tol=1e-6}}}

\item{iter.max}{Maximum Iteration number, default is \strong{\code{iter.max=20L}}}

\item{method}{Selecting Method used, default is \strong{\code{method="ProxN"}}}

\item{lambda}{Parameter for Proximal Newton's Method. Default is \strong{\code{lambda=1e8}}}

\item{btr}{Backtracking line search approach, default is \strong{\code{btr="dynamic"}}:}

\item{tau}{In the Newton's Method, Default is \strong{\code{tau=0.5}}. Used to control for step size.}

\item{stop}{Stopping rule, default is \strong{\code{stop="ratch"}}.}

\item{parallel}{Parallel computation, Default is \strong{\code{parallel=FALSE}}}

\item{threads}{Parallel computation parameter(number of cores)Default is \strong{\code{threads=1L}}}

\item{degree}{Degree of smoothing spline. Default setting is \strong{\code{degree=3L}}.}

\item{TIC_prox}{When calculating information criteria, there might be numerical issue(second order derivative,
Hessian matrix approximate is singular), thus we proposed to add a small term to the diagonal. Default \strong{\code{TIC_prox = FALSE}}}

\item{lambda.spline}{Smoothing parameter lambda. Default is \strong{\code{lambda.spline = 0}} which refers to Newton's Method without penalization.}

\item{ord}{Specify which derivative to penalize. Default setting is \strong{\code{ord=4}}.}

\item{fixedstep}{There might be times when the stopping criteria not working, thus,
the number of steps could be set manually. Default value is \strong{\code{fixedstep = FALSE}}, if it is true, will stop by \code{iter.max}}

\item{ICLastOnly}{Only calculate the last information criteria if is TRUE. Default is \strong{\code{ICLastOnly=FALSE}}}
}
\value{
\item{model.cv}{a \code{"coxtp"} object with penalization coefficient chosen based on cross validation}
\item{lambda.selected}{The selected penalization coefficients for different information criteria}
}
\description{
Fit a cox Non-proportional Hazards model via penalized maximum likelihood.
}
\examples{
data(ExampleData)
z <- ExampleData$x
time  <- ExampleData$time
event <- ExampleData$event

lambda.spline  = c(0,1)
fit   <- ic.coxtp(event = event, z = z, time = time, lambda.spline=lambda.spline)
}
