% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ExampleData}
\alias{ExampleData}
\title{Example data for surtvep with 4000 observations of 2 variables:}
\format{
A List containing the following elements:
\describe{
\item{x}{Simulated covariate V1 and v2, Continuous variable. True time-independent function is b(t)=1 and true time-dependent function is b(t)=sin(3\emph{pi}t/4)}
\item{event}{Simulated event variable, Binary varible with 0, 1 }
\item{time}{Simulated time variable, Continous variable with non-negative value}
}
}
\usage{
data(ExampleData)
}
\description{
A simulated data set containing two 2 variables.
}
\keyword{datasets}
