% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ExampleData}
\alias{ExampleData}
\title{example data with 2000 observations of 2 continuous variables}
\format{
A list containing the following elements:
\describe{
\item{z}{simulated continuous covariates V1 and V2, with a time-independent coefficient \eqn{\beta_1(t)=1}
and a time-varying coefficient \eqn{\beta_2(t)=sin(3\pi t/4).}}
\item{event}{simulated failure event response; binary variable with 0 or 1.}
\item{time}{simulated observed event times; continuous variable with non-negative values.}
}
}
\usage{
data(ExampleData)
}
\description{
A simulated data set containing 2 continuous variables.
}
\keyword{datasets}
