% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ExampleData}
\alias{ExampleData}
\title{example data for surtvep with 2000 observations of 2 variables}
\format{
A list containing the following elements:
\describe{
\item{x}{simulated continuous covariate V1 and V2. The true time-independent function is b(t)=1 and the true time-dependent function is b(t)=sin(3\emph{pi}t/4).}
\item{event}{simulated event variable, binary variable with 0, 1.}
\item{time}{simulated time variable, continuous variable with non-negative value.}
}
}
\usage{
data(ExampleData)
}
\description{
A simulated data set containing two variables.
}
\keyword{datasets}
