% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ExampleDataBinary}
\alias{ExampleDataBinary}
\title{example data with 2000 observations of 2 binary variables}
\format{
A list containing the following elements:
\describe{
\item{z}{simulated binary covariates V1 and V2, with a time-independent coefficient \eqn{\beta_1(t)=1}
and a time-varying coefficient \eqn{\beta_2(t)=exp(-1.5t).}}
\item{event}{simulated failure event response; binary variable with 0 or 1.}
\item{time}{simulated observed event times; continuous variable with non-negative values. }
}
}
\usage{
data(ExampleDataBinary)
}
\description{
A simulated data set containing 2 binary variables.
}
\keyword{datasets}
