% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ExampleDataBinary}
\alias{ExampleDataBinary}
\title{example data for surtvep with 2000 observations of 2 binary variables}
\format{
a list containing the following elements:
\describe{
\item{V1}{simulated binary covariate V1. True time-dependent function is b(t)=1. }
\item{V2}{simulated binary covariate V2. True time-dependent function is b(t)=exp(-1.5*t). }
\item{event}{simulated event variable, binary variable with 0, 1. }
\item{time}{simulated time variable, continuous variable with non-negative value. }
}
}
\usage{
data(ExampleDataBinary)
}
\description{
A simulated data set containing two 2 variables.
}
\keyword{datasets}
