% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.tvcoef.R
\name{get.tvcoef}
\alias{get.tvcoef}
\title{helper function to get time-varying coefficients}
\usage{
get.tvcoef(fit, times)
}
\arguments{
\item{fit}{model from \code{coxtv} or \code{coxtp}.}

\item{times}{the observation time. If \code{NULL}, the observation time for fitting the model will be used.}
}
\value{
a matrix of the time-varying coefficients. The dimension is \code{length(times)} x \code{nvars}, where \code{nvars} is the number
of covariates in the fitted mode.
Each row represents the time-varying coefficients at the corresponding time.
``
}
\description{
Gives the time-varying coefficients based on a fitted \code{coxtv} or \code{coxtp} subject. Users can specify  the observation  time.
}
\examples{
z     <- ExampleData$x
time  <- ExampleData$time
event <- ExampleData$event
fit   <- coxtv(event = event, z = z, time = time, degree = 2)
coef  <- get.tvcoef(fit)


}
