% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{plot.baseline}
\alias{plot.baseline}
\title{plotting baseline}
\usage{
\method{plot}{baseline}(fit, xlab, ylab, xlim, ylim, title, ...)
}
\arguments{
\item{fit}{model get from \code{baseline} function.}
}
\description{
Plotting baseline from a fitted \code{baseline} object.
}
\examples{
\dontrun{
data(ExampleData)
z <- ExampleData$x
time  <- ExampleData$time
event <- ExampleData$event

fit   <- coxtv(event = event, z = z, time = time)
base.est <- baseline(fit)
}
}
