% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{plot.baseline}
\alias{plot.baseline}
\title{plotting baseline from a fitted \code{coxtp.baseline} object}
\usage{
\method{plot}{baseline}(fit, xlab, ylab, xlim, ylim, title, ...)
}
\arguments{
\item{fit}{Model get from coxtp.baseline}
}
\description{
plotting baseline from a fitted \code{coxtp.baseline} object
}
\examples{
\dontrun{
data(ExampleData)
z <- ExampleData$x
time  <- ExampleData$time
event <- ExampleData$event

fit   <- coxtv(event = event, z = z, time = time)
base.est <- baseline(fit)
}
}
