% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtv.plot.R
\name{plot.coxtv}
\alias{plot.coxtv}
\title{plotting result from a \code{coxtv} object}
\usage{
\method{plot}{coxtv}(
  fit,
  times,
  parm,
  CI = TRUE,
  level = 0.95,
  exponentiate = FALSE,
  xlab,
  ylab,
  xlim,
  ylim,
  save = FALSE,
  allinone = FALSE,
  title,
  linetype,
  fill,
  color,
  labels,
  expand,
  ...
)
}
\arguments{
\item{fit}{Model get from coxtv}
}
\description{
plotting result from a \code{coxtv} object
}
\examples{
data(ExampleData)
z <- ExampleData$x
time <- ExampleData$time
event <- ExampleData$event
fit <- coxtv(event = event, z = z, time = time)
plot(fit)

}
