% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{StrataExample}
\alias{StrataExample}
\title{example data for stratified model illustration}
\format{
A list containing the following elements:
\describe{
\item{z}{simulated binary covariates V1 and V2, with a time-independent coefficient \eqn{\beta_1(t)=1}
and a time-varying coefficient \eqn{\beta_2(t)=sin(3\pi t/4).}}
\item{event}{simulated failure event response; binary variable with 0 or 1.}
\item{time}{simulated observed event times; continuous variable with non-negative values. }
\item{strata}{simulated strata variable; patients in different stratums have different baseline hazards.}
}
}
\usage{
data(StrataExample)
}
\description{
A simulated data set containing 2 binary variables from 10 distinct stratums.
}
\keyword{datasets}
