% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{StrataExample}
\alias{StrataExample}
\title{Example data for stratified model illustration}
\format{
A List containing the following elements:
\describe{
\item{x}{Simulated covariate V1 and v2, Continuous variable. True time-dependent function is b(t)=1 and true time-dependent function is b(t)=sin(3\emph{pi}t/4)}
\item{event}{Simulated event variable, Binary varible with 0, 1 }
\item{time}{Simulated time variable, Continous variable with non-negative value}
\item{strata}{Simulated strata variable, patients in different stratums have different baseline hazards}
}
}
\usage{
data(StrataExample)
}
\description{
A simulated data set containing two 2 variables.
}
\keyword{datasets}
