% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvef.ph.R
\name{tvef.zero}
\alias{tvef.zero}
\title{test the significance of the covariates from a \code{coxtv} or \code{coxtp} object}
\usage{
tvef.zero(fit, parm)
}
\arguments{
\item{fit}{fitted \code{"coxtv"} or \code{"coxtp"}  model.}

\item{parm}{the names of parameter to be tested.}
}
\value{
\code{tvef.zero} produces a matrix. Each row corresponds to a covariate from the fitted object. The three
columns gives the value of test statistic, degrees of freedom and P-value.
}
\description{
Test the significance of the covariates from a \code{coxtv} or \code{coxtp} object using a Wald test statistic.
}
\examples{
data(ExampleData)
z <- ExampleData$x
time <- ExampleData$time
event <- ExampleData$event
fit <- coxtv(event = event, z = z, time = time)
tvef.ph(fit)

}
\seealso{
\code{\link{tvef.ph}} \code{\link{tvef.zero.time}}
}
