% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvef.ph.R
\name{tvef.zero}
\alias{tvef.zero}
\title{test the significance of the covariates from a \code{coxtv} or \code{coxtp} object}
\usage{
tvef.zero(fit, times, parm)
}
\arguments{
\item{fit}{fitted \code{"coxtv"} or \code{"coxtp"}  model}

\item{times}{the time points to test if the covariate is significant or not}

\item{parm}{the names of parameter to be tested}
}
\description{
test the significance of the covariates from a \code{coxtv} or \code{coxtp} object
}
\examples{
data(ExampleData)
z <- ExampleData$x
time <- ExampleData$time
event <- ExampleData$event
fit <- coxtv(event = event, z = z, time = time)
tvef.zero(fit)

}
